som-external: Support secure boot images without secure rootfs
The need emerged to have ability to completely re-flash software
and provision secure file system with custom data on the modules
with secure boot enabled.

To accomplish this we need to be able to create SD card images
with secure boot and secure file system key, but without security
on rootfs.

Bug 23991