lrd-encrypted-storage-toolkit: Add support for dropbear SSH service
Add dropbear to services linked to /data/secret. This allows the dropbear
service to save keys instead of regenerating them every boot.
Dropbear is setup to use a symlink to handle its checks for RO/RW
filesystems. We want to overwrite that symlink and point to /data/secret.
Add dropbear@.service overrides to remove RO/RW check and inherit the
key ring.

Bug 23554