erase_data.sh: dont use cp -t, older busybox doesnt support it
Older versions of busybox does not support the -t option. This causes an
error when updating from an older version of busybox as it cant handle -t.
erase_data.sh is included in .swu and it needs to maintain compatibility
across a wide range of deployments. Revert to using the previous cp command.

Bug 24989